pyproject.toml
[project]
name = "python-project-template"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=0.3.7",
    "langgraph-supervisor>=0.0.8",
    "langgraph-swarm>=0.0.5",
    "python-dotenv>=1.0.1",
    "trustcall>=0.0.38",
]

[tool.mypy]
strict = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "TRY", "EM", "FA", "SIM108", "S101", "RET505", "TCH002"]
